data-prepper-pipeline-parser/src/test/java/org/opensearch/dataprepper/pipeline/parser/rule/RuleParserTest.java
//package org.opensearch.dataprepper.pipeline.parser.rule;
//
//import static org.mockito.Mockito.*;
//        import static org.junit.jupiter.api.Assertions.*;
//
//        import org.junit.jupiter.api.Test;
//        import org.mockito.InjectMocks;
//        import org.mockito.Mock;
//        import com.fasterxml.jackson.databind.ObjectMapper;
//
//class RuleParserTest {
//
//    @Mock
//    ObjectMapper objectMapper;
//
//    @InjectMocks
//    RuleParser ruleParser;
//
//    @Test
//    void testParseRuleFileSuccess() throws Exception {
//        // Assuming a simple RuleTransformerModel for demonstration
//        RuleTransformerModel ruleModel = new RuleTransformerModel();
//        ruleModel.setJsonPath("$..somePath");
//
//        String ruleJson = "{\"json_path\": \"$..somePath\"}";
//        when(objectMapper.readValue(anyString(), eq(RuleTransformerModel.class))).thenReturn(ruleModel);
//
//        ruleParser.parseRuleFile("path/to/rule-file");
//
//        // Verify the ObjectMapper was called with the expected JSON
//        verify(objectMapper).readValue(ruleJson, RuleTransformerModel.class);
//
//        // Assert that the rule was added successfully
//        assertFalse(ruleParser.getRulesJsonPath().isEmpty());
//        assertEquals("$..somePath", ruleParser.getRulesJsonPath().get(0));
//    }
//
//    // Additional setup necessary for Mockito annotations
//}
